+++ houston/ardu/__init__.py
@@ -1,5 +1,3 @@
-import houston.ardu.base
-import houston.ardu.sandbox
-from houston.ardu.copter import ArduCopter
-from houston.ardu.rover import ArduRover
-# from plane import ArduPlane
+from . import sandbox, base
+from .copter import ArduCopter
+from .rover import ArduRover
