+++ houston/ardu/common.py
@@ -1,16 +1,18 @@
-__all__ = ['ArmDisarmSchema']
+__all__ = ['GotoLoiter', 'ArmDisarmSchema']
 
 import time
 
+import geopy
+import geopy.distance
 from pymavlink import mavutil
 
-from ...configuration import Configuration
-from ...command import CommandSchema, Parameter, Command
-from ...specification import Specification, Idle
-from ...state import State
-from ...environment import Environment
-from ...valueRange import DiscreteValueRange
-from ..sandbox import Sandbox
+from .sandbox import Sandbox
+from ..state import State
+from ..environment import Environment
+from ..valueRange import ContinuousValueRange, DiscreteValueRange
+from ..configuration import Configuration
+from ..command import CommandSchema, Parameter, Command
+from ..specification import Specification, Idle
 
 
 class ArmDisarmSchema(CommandSchema):
@@ -65,3 +67,26 @@ class DisarmNormally(Specification):
         post = lambda a, s0, s1, e, c: not s1.armed
         timeout = lambda a, s, e, c: c.constant_timeout_offset + 1.0
         super().__init__('disarm-normal', pre, post, timeout)
+
+
+class GotoLoiter(Specification):
+    """
+    If the robot is armed and in its `LOITER` mode, GoTo actions should have no
+    effect upon the robot. (Why isn't this covered by Idle?)
+    """
+    def __init__(self) -> None:
+        pre = lambda a, s, e, c: s.armed and s.mode == 'LOITER'
+        timeout = lambda a, s, e, c: c.constant_timeout_offset
+
+        def post(a, s0, s1, e, c) -> bool:
+            sat_mode = s1.mode == 'LOITER'
+            # FIXME 82
+            err_lon = 0.1
+            err_lat = 0.1
+            err_alt = 0.5
+            sat_lon = abs(s1.longitude - s0.longitude) <= err_lon
+            sat_lat = abs(s1.latitude - s0.latitude) <= err_lat
+            sat_alt = abs(s1.altitude - s0.altitude) <= err_alt
+            return sat_mode and sat_lon and sat_lat and sat_alt
+
+        super().__init__('loiter', pre, post, timeout)
