--- houston/ardu/common/__init__.py
@@ -1,1 +1,0 @@
-from houston.ardu.common.arm import ArmDisarmSchema
